Removed Invoice.invoice_id (in later phase Invoice.uid) from two reasons   * It was dependent on PK thus it could be computed after save which is inpractical for     those who use signals   * In many countries the identification symbol of invoice has to be a number. Why to     duplicate Invoice.PK which is natural key?
Added some testing data

Tricky part with translation of currency units and date format. Default is U.S. values
and they are within gettext function. Everyone is welcome to add their own locale. I think
this is better solution than to introduce milions of settings which will get duplicated for
every general app which are used in the project. Btw. the currency format is not just _("$")
it's whole formating _("{0} $") so feel free to make a translation like _("& {0:.2}").